requirements.txt
psycopg2-binary
pgvector
requests
beautifulsoup4